add interpolation algorithm by calculating on CPU

+++ cuda/binarization/binarize_kernel.cu
@@ -2,8 +2,8 @@
 
 __global__ void binarizeKernel(short * volumedata, short * binarydata, int size, int thresh, int maxval)
 {
-	int blockId = blockIdx.x + blockIdx.y * gridDim.x + gridDim.x * gridDim.y * blockIdx.z;
-	int threadId = blockId * blockDim.x + threadIdx.x;
+	unsigned long blockId = blockIdx.x + blockIdx.y * gridDim.x + gridDim.x * gridDim.y * blockIdx.z;
+	unsigned long threadId = blockId * blockDim.x + threadIdx.x;
 
 	if (volumedata[threadId] < thresh)
 	{
